refactoring simulator to use spock. more examples running
