+++ examples/parametric_mab.py
@@ -3,7 +3,12 @@
 import pandas as pd
 from sklearn.preprocessing import StandardScaler
 
-from mabwiser.mab import MAB, LearningPolicy, NeighborhoodPolicy
+from mabwiser.mab import MAB
+
+from mabwiser.configs.mab import MABConfig
+from mabwiser.configs.arm import ArmConfig, WarmStartConfig
+from mabwiser.configs.learning import LinUCB, LinTS
+from mabwiser.configs.neighborhood import Radius, KNearest
 
 ######################################################################################
 #
@@ -19,10 +24,10 @@ from mabwiser.mab import MAB, LearningPolicy, NeighborhoodPolicy
 ######################################################################################
 
 # Arms
-ads = [1, 2, 3, 4, 5]
+ads = ["1", "2", "3", "4", "5"]
 
 # Historical data of ad decisions with corresponding revenues and context information
-train_df = pd.DataFrame({'ad': [1, 1, 1, 2, 4, 5, 3, 3, 2, 1, 4, 5, 3, 2, 5],
+train_df = pd.DataFrame({'ad': ["1", "1", "1", "2", "4", "5", "3", "3", "2", "1", "4", "5", "3", "2", "5"],
                          'revenues': [10, 17, 22, 9, 4, 20, 7, 8, 20, 9, 50, 5, 7, 12, 10],
                          'age': [22, 27, 39, 48, 21, 20, 19, 37, 52, 26, 18, 42, 55, 57, 38],
                          'click_rate': [0.2, 0.6, 0.99, 0.68, 0.15, 0.23, 0.75, 0.17,
@@ -31,7 +36,7 @@ train_df = pd.DataFrame({'ad': [1, 1, 1, 2, 4, 5, 3, 3, 2, 1, 4, 5, 3, 2, 5],
                         )
 
 # Arm features for warm start
-arm_to_features = {1: [0, 1, 1], 2: [0, 0.5, 0.5], 3: [1, 1, 0.5], 4: [0.2, 1, 0], 5: [0, 1, 0.1], 6: [0, 0.5, 0.5]}
+arm_to_features = {"1": [0.0, 1.0, 1.0], "2": [0.0, 0.5, 0.5], "3": [1.0, 1.0, 0.5], "4": [0.2, 1.0, 0.0], "5": [0.0, 1.0, 0.1], "6": [0.0, 0.5, 0.5]}
 
 # Test data to for new prediction
 test_df = pd.DataFrame({'age': [37, 52], 'click_rate': [0.5, 0.6], 'subscriber': [0, 1]})
@@ -47,8 +52,11 @@ test = scaler.transform(test_df.values.astype('float64'))
 ##################################################
 
 # LinUCB learning policy with alpha 1.25 and l2_lambda 1
-linucb = MAB(arms=ads,
-             learning_policy=LearningPolicy.LinUCB(alpha=1.25, l2_lambda=1))
+linucb = MAB(config=MABConfig(
+    arms=ads,
+    learning_policy=LinUCB(alpha=1.25, l2_lambda=1.0),
+    seed=123456
+))
 
 # Learn from previous ads shown and revenues generated
 linucb.fit(decisions=train_df['ad'], rewards=train_df['revenues'], contexts=train)
@@ -61,25 +69,28 @@ expectations = linucb.predict_expectations(test)
 
 # Results
 print("LinUCB: ", prediction, " ", expectations)
-assert(prediction == [5, 2])
+assert(prediction == ["5", "2"])
 
 # Online update of model
 linucb.partial_fit(decisions=prediction, rewards=test_df_revenue, contexts=test)
 
 # Update the model with new arm
-linucb.add_arm(6)
+linucb.add_arm(ArmConfig(arm="6"))
 
 # Warm start new arm
-linucb.warm_start(arm_to_features, distance_quantile=0.75)
+linucb.warm_start(WarmStartConfig(arm_to_features=arm_to_features, distance_quantile=0.75))
 
 ###################################################################
 # LinUCB Learning Policy combined with Radius Neighborhood Policy
 ###################################################################
 
 # Radius context policy with radius equals to 1 and LinUCB learning with alpha of 1
-radius = MAB(arms=ads,
-             learning_policy=LearningPolicy.LinUCB(alpha=1),
-             neighborhood_policy=NeighborhoodPolicy.Radius(radius=1))
+radius = MAB(config=MABConfig(
+    arms=ads,
+    learning_policy=LinUCB(alpha=1.0),
+    neighborhood_policy=Radius(radius=1.0),
+    seed=123456
+))
 
 # Learn from previous ads shown and revenues generated
 radius.fit(decisions=train_df['ad'], rewards=train_df['revenues'], contexts=train)
@@ -92,16 +103,19 @@ expectations = radius.predict_expectations(test)
 
 # Results
 print("Radius: ", prediction, " ", expectations)
-assert(prediction == [1, 2])
+assert(prediction == ["1", "2"])
 
 #####################################################################
 # LinUCB Learning Policy combined with KNearest Neighborhood Policy
 #####################################################################
 
 # KNearest context policy with k equals to 4 and LinUCB learning with alpha of 1.25
-knearest = MAB(arms=ads,
-               learning_policy=LearningPolicy.LinUCB(alpha=1.25),
-               neighborhood_policy=NeighborhoodPolicy.KNearest(k=4))
+knearest = MAB(config=MABConfig(
+    arms=ads,
+    learning_policy=LinUCB(alpha=1.0),
+    neighborhood_policy=KNearest(k=4),
+    seed=123456
+))
 
 # Learn from previous ads shown and revenues generated
 knearest.fit(decisions=train_df['ad'], rewards=train_df['revenues'], contexts=train)
@@ -114,15 +128,17 @@ expectations = knearest.predict_expectations(test)
 
 # Results
 print("KNearest: ", prediction, " ", expectations)
-assert(prediction == [1, 2])
+assert(prediction == ["1", "2"])
 
 ##################################################
 # Linear Thompson Sampling Learning Policy
 ##################################################
 
 # LinTS learning policy with alpha 1.25 and l2_lambda 1
-lints = MAB(arms=ads,
-            learning_policy=LearningPolicy.LinTS(alpha=1.5, l2_lambda=1))
+lints = MAB(config=MABConfig(
+    arms=ads,
+    learning_policy=LinTS(alpha=1.5, l2_lambda=1.0),
+))
 
 # Learn from previous ads shown and revenues generated
 lints.fit(decisions=train_df['ad'], rewards=train_df['revenues'], contexts=train)
@@ -135,22 +151,24 @@ expectations = lints.predict_expectations(test)
 
 # Results
 print("LinTS: ", prediction, " ", expectations)
-assert(prediction == [5, 2])
+assert(prediction == ["5", "2"])
 
 # Online update of model
 lints.partial_fit(decisions=prediction, rewards=test_df_revenue, contexts=test)
 
 # Update the model with new arm
-lints.add_arm(6)
+lints.add_arm(ArmConfig(arm="6"))
 
 ###################################################################
 # LinTS Learning Policy combined with Radius Neighborhood Policy
 ###################################################################
 
 # Radius context policy with radius equals to 1 and LinTS learning with alpha of 1
-radius = MAB(arms=ads,
-             learning_policy=LearningPolicy.LinTS(alpha=0.5),
-             neighborhood_policy=NeighborhoodPolicy.Radius(radius=1))
+radius = MAB(config=MABConfig(
+    arms=ads,
+    learning_policy=LinTS(alpha=0.5),
+    neighborhood_policy=Radius(radius=1.0)
+))
 
 # Learn from previous ads shown and revenues generated
 radius.fit(decisions=train_df['ad'], rewards=train_df['revenues'], contexts=train)
@@ -163,16 +181,18 @@ expectations = radius.predict_expectations(test)
 
 # Results
 print("Radius: ", prediction, " ", expectations)
-assert(prediction == [1, 2])
+assert(prediction == ["1", "2"])
 
 #####################################################################
 # LinTS Learning Policy combined with KNearest Neighborhood Policy
 #####################################################################
 
 # KNearest context policy with k equals to 4 and LinTS learning with alpha of 1.25
-knearest = MAB(arms=ads,
-               learning_policy=LearningPolicy.LinTS(alpha=1),
-               neighborhood_policy=NeighborhoodPolicy.KNearest(k=4))
+knearest = MAB(config=MABConfig(
+    arms=ads,
+    learning_policy=LinTS(alpha=1.0),
+    neighborhood_policy=KNearest(k=4)
+))
 
 # Learn from previous ads shown and revenues generated
 knearest.fit(decisions=train_df['ad'], rewards=train_df['revenues'], contexts=train)
@@ -185,4 +205,4 @@ expectations = knearest.predict_expectations(test)
 
 # Results
 print("KNearest: ", prediction, " ", expectations)
-assert(prediction == [1, 2])
+assert(prediction == ["1", "2"])
