+++ mabwiser/utilities/general.py
@@ -7,6 +7,7 @@ This module provides various general utilities
 
 import multiprocessing as mp
 from typing import Dict
+import numpy as np
 
 from mabwiser.utilities.types import Num
 
@@ -37,3 +38,39 @@ def reset(dictionary: Dict, value) -> None:
     Maps every key to the given value.
     """
     dictionary.update({}.fromkeys(dictionary, value))
+
+
+def get_stats(rewards: np.ndarray) -> Dict:
+    """Calculates descriptive statistics for the given array of rewards.
+
+    Parameters
+    ----------
+    rewards: nd.nparray
+        Array of rewards for a single arm.
+
+    Returns
+    -------
+    A dictionary of descriptive statistics.
+    Dictionary has the format {'count', 'sum', 'min', 'max', 'mean', 'std'}
+    """
+    return {
+        "count": rewards.size,
+        "sum": rewards.sum(),
+        "min": rewards.min(),
+        "max": rewards.max(),
+        "mean": rewards.mean(),
+        "std": rewards.std(),
+    }
+
+
+def get_context_hash(contexts, plane):
+    # Project rows onto plane and get signs
+    projection_signs = 1 * (np.dot(contexts, plane) > 0)
+
+    # Get base 2 value of projection signs
+    # Another approach is to convert to strings ('01000', '00101', '11111', etc)
+    hash_values = np.zeros(contexts.shape[0])
+    for i in range(plane.shape[1]):
+        hash_values = hash_values + (projection_signs[:, i] * 2**i)
+
+    return hash_values
